+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 [Unreleased](https://github.com/jshwi/pyaud/compare/v4.1.0...HEAD)
 ------------------------------------------------------------------------
+### Added
+- Add `--exclude` argument
+
 ### Changed
 - Change `audit` arg type
 - File exclusions now uses regex instead of list